backend/requirements.txt
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.8.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
db==0.1.1
db-sqlite3==0.0.1
exceptiongroup==1.2.2
fastapi==0.115.7
Flask==3.1.0
greenlet==3.1.1
h11==0.14.0
helpers==0.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
passlib==1.7.4
prometheus-fastapi-instrumentator==7.0.2
prometheus_client==0.21.1
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
